Improving the code to read the vtk file faster, and providing the buttons and its callbacks to show the flow computation workflow

+++ BoundaryController.cpp
@@ -54,7 +54,7 @@ QRectF BoundaryController::boundingRect() const
 {
     QPainterPath path;
 
-    int ncurves = boundary.size();
+    int ncurves = (int) boundary.size();
     for( int i = 0; i < ncurves; ++i )
     {
         QPainterPath curve = boundary.at( i );
